ArchaeoChron: Bayesian Modeling of Archaeological Chronologies

Provides a list of functions for the Bayesian modeling of archaeological chronologies. The Bayesian models are implemented in 'JAGS' (Plummer 2003). The inputs are measurements with their associated standard deviations and the study period. The output is the MCMC sample of the posterior distribution of the event date with or without radiocarbon calibration.

Details

**Version**	0.2

**License**	GPL-3

**CRAN DOI**	doi:10.32614/CRAN.package.ArchaeoChron

Laboratoire de Mathématiques Jean Leray (UMR 6629)
2, rue de la Houssinière
BP 92208
F-44322 Nantes Cedex 3
France

See also

Author

Maintainer: Anne Philippe anne.philippe@univ-nantes.fr (ORCID)

Authors:

Marie-Anne Vibet (ORCID)

Other contributors:

Nicolas Frerebeau nicolas.frerebeau@u-bordeaux-montaigne.fr (ORCID) [contributor]

Nantes Université (ROR) [funder]

CNRS (ROR) [funder]