Combine Gaussian Dates with a Random Effect

Combine Gaussian Dates with a Random Effect

combinationWithRandomEffect_Gauss(
  M,
  s,
  studyPeriodMin,
  studyPeriodMax,
  refYear = NULL,
  numberChains = 2,
  numberAdapt = 10000,
  numberUpdate = 10000,
  variable.names = c("theta"),
  numberSample = 50000,
  thin = 10
)

Arguments

M

A [`numeric`] vector of measurements.

s

A [`numeric`] vector of errors.

studyPeriodMin

A length-one [`numeric`] vector specifying the start time of the study period.

studyPeriodMax

A length-one [`numeric`] vector specifying the end time of the study period.

refYear

A [`numeric`] vector specifying the reference year. If `NULL` (the default), AD years are expected.

numberChains

An [`integer`] giving the number of of parallel chains for the model (see [jags.model()]).

numberAdapt

An [`integer`] giving the number of iterations for adaptation (see [jags.model()]).

numberUpdate

An [`integer`] giving the number of iterations to update the model by.

variable.names

A [`character`] vector giving the names of variables to be monitored (see [coda.samples()]).

numberSample

An [`integer`] giving the number of iterations to monitor (see [coda.samples()]).

thin

An [`integer`] giving the thinning interval for monitors (see [coda.samples()]).

Value

An [`mcmc.list`][coda::mcmc.list()] object.

Author

A. Philippe, M.-A. Vibet