ArchaeoChron: Bayesian Modeling of Archaeological Chronologies

Provides a list of functions for the Bayesian modeling of archaeological chronologies. The Bayesian models are implemented in 'JAGS' ('JAGS' stands for Just Another Gibbs Sampler. It is a program for the analysis of Bayesian hierarchical models using Markov Chain Monte Carlo (MCMC) simulation. See http://mcmc-jags.sourceforge.net/ and "JAGS Version 4.3.0 user manual", Martin Plummer (2017) https://sourceforge.net/projects/mcmc-jags/files/Manuals/.). The inputs are measurements with their associated standard deviations and the study period. The output is the MCMC sample of the posterior distribution of the event date with or without radiocarbon calibration.

Details

**Package:**	ArchaeoChron

**Type:**	Package

**Version:**	0.2

**License:**	GPL-3

See also

Author

**Full list of authors and contributors** (alphabetic order)

Nicolas Frerebeau	*Université Bordeaux Montaigne, France*

Anne Philippe	*Université de Nantes, France*

Marie-Anne Vibet	*Université de Nantes, France*

**Package maintainer**

Anne Philippe
anne.philippe@univ-nantes.fr

Laboratoire de Mathématiques Jean Leray (UMR 6629)
2, rue de la Houssinière
BP 92208
F-44322 Nantes Cedex 3
France